+++ R/counts.R
@@ -52,16 +52,14 @@ countcorrect <- function(data, .by = NULL, .input = NULL, .extra = NULL) {
         unnest(all_of(.input$name_acc))
     }
     data <- data |>
-      group_by(pick(all_of(.by))) |>
       summarise(
         # `NA` might be produced in parsing characters and should be ignored
         "{.input$name_nc}" := sum(.data[[.input$name_acc]] == 1, na.rm = TRUE),
-        .groups = "drop"
+        .by = all_of(.by)
       )
   }
   data |>
-    group_by(pick(all_of(.by))) |>
-    summarise(nc = sum(.data[[.input$name_nc]]), .groups = "drop") |>
+    summarise(nc = sum(.data[[.input$name_nc]]), .by = all_of(.by)) |>
     vctrs::vec_restore(data)
 }
 
@@ -72,20 +70,18 @@ countcorrect2 <- function(data, .by = NULL, .input = NULL, .extra = NULL) {
     update_settings(.input)
   if (!all(has_name(data, .input[c("name_nc", "name_ne")]))) {
     data <- data |>
-      group_by(pick(all_of(.by))) |>
       summarise(
         "{.input$name_nc}" := sum(.data[[.input$name_acc]] == 1),
         "{.input$name_ne}" := sum(.data[[.input$name_acc]] == 0),
-        .groups = "drop"
+        .by = all_of(.by)
       )
   }
   data |>
-    group_by(pick(all_of(.by))) |>
     summarise(
       nc_cor = sum(
         .data[[.input$name_nc]] - .data[[.input$name_ne]]
       ),
-      .groups = "drop"
+      .by = all_of(.by)
     ) |>
     vctrs::vec_restore(data)
 }
@@ -96,13 +92,12 @@ sumweighted <- function(data, .by = NULL, .input = NULL, .extra = NULL) {
   .input <- list(name_weight = "nstim", name_acc = "acc") |>
     update_settings(.input)
   data |>
-    group_by(pick(all_of(.by))) |>
     summarise(
       nc_weighted = sum(
         .data[[.input$name_weight]] *
           (.data[[.input$name_acc]] == 1)
       ),
-      .groups = "drop"
+      .by = all_of(.by)
     ) |>
     vctrs::vec_restore(data)
 }
@@ -113,10 +108,9 @@ sumscore <- function(data, .by = NULL, .input = NULL, .extra = NULL) {
   .input <- list(name_score = "score") |>
     update_settings(.input)
   data |>
-    group_by(pick(all_of(.by))) |>
     summarise(
       nc_score = sum(as.numeric(.data[[.input$name_score]])),
-      .groups = "drop"
+      .by = all_of(.by)
     ) |>
     vctrs::vec_restore(data)
 }