+++ tests/testthat/test-capacity.R
@@ -7,7 +7,6 @@ test_that("Works without grouping variables", {
   ) |>
     mutate(acc = sample(c(0, 1), n(), replace = TRUE))
   capacity(data) |>
-    expect_silent() |>
     expect_snapshot_value(style = "json2")
 })
 
@@ -19,7 +18,6 @@ test_that("Return nan values when false alarm is 1", {
   ) |>
     mutate(acc = c(rep(0, 20), rep(1, n() - 20)))
   capacity(data) |>
-    expect_silent() |>
     purrr::pluck("k") |>
     is.nan() |>
     expect_true()
@@ -35,6 +33,5 @@ test_that("Works with grouping variables", {
   ) |>
     mutate(acc = sample(c(0, 1), n(), replace = TRUE))
   capacity(data, .by = "id") |>
-    expect_silent() |>
     expect_snapshot_value(style = "json2")
 })