+++ NEWS.md
@@ -1,3 +1,15 @@
+# preproc.iquizoo 2.6.3
+
+## Breaking Changes
+
+* Added `"cutoff"` method in response time outlier detection, which is used in `srt()` now.
+* Added `"transform"` method in response time outlier detection, which was stated to be used but the code was not right. Now it is used as default if not specified.
+
+## Miscellaneous
+
+* Let internal function `calc_sdt()` throw error when type column is not valid.
+* Enhanced code quality of internal functions.
+
 # preproc.iquizoo 2.6.2
 
 ## Bug Fixes